+++ src/commons/linear_systems.jl
@@ -103,10 +103,10 @@ either providing a tangent space or a manifold and a point.
 This can be evaluated in-place of `Y`.
 """
 
-@doc "$(_doc_get_vector_field_slso)"
 function get_vector_field(M::AbstractManifold, admo::AbstractDecoratedManifoldObjective, p)
     return get_vector_field(M, get_objective(admo, false), p)
 end
+@doc "$(_doc_get_vector_field_slso)"
 function get_vector_field(M::AbstractManifold, slso::SymmetricLinearSystemObjective, p)
     Y = zero_vector(M, p)
     return slso.b!(M, Y, p)
@@ -118,20 +118,20 @@ function get_vector_field!(M::AbstractManifold, Y, slso::SymmetricLinearSystemOb
     return slso.b!(M, Y, p)
 end
 # Also on TpM – shortcuts
-@doc "$(_doc_get_vector_field_slso)"
 function get_vector_field(TpM::TangentSpace, admo::AbstractDecoratedManifoldObjective)
     return get_vector_field(TpM, get_objective(admo, false))
 end
+@doc "$(_doc_get_vector_field_slso)"
 function get_vector_field(TpM::TangentSpace, slso::SymmetricLinearSystemObjective)
     M = base_manifold(TpM)
     p = base_point(TpM)
     Y = zero_vector(M, p)
     return slso.b!(M, Y, p)
 end
-@doc "$(_doc_get_vector_field_slso)"
 function get_vector_field!(TpM::TangentSpace, Y, admo::AbstractDecoratedManifoldObjective)
     return get_vector_field!(TpM, Y, get_objective(admo, false))
 end
+@doc "$(_doc_get_vector_field_slso)"
 function get_vector_field!(TpM::TangentSpace, Y, slso::SymmetricLinearSystemObjective)
     M = base_manifold(TpM)
     p = base_point(TpM)